#44 extracted common component, more smooth adjusting of movement factor
