+++ client/client/style/main.css
@@ -299,11 +299,11 @@ background: #00c4c6;
     border-radius: 50px;
 }
 
-.camera-controls {
+.adjuster-controls {
     display: flex;
     flex-direction: row;
 }
-.camera-controls button {
+.adjuster-controls button {
     transition: all .2s;
     border: none;
     background: cyan;
@@ -317,10 +317,10 @@ background: #00c4c6;
     box-shadow: 1px 1px 1px black;
     margin-bottom: 10px;
 }
-.camera-controls button:hover {
+.adjuster-controls button:hover {
     background: darkcyan;
 }
-.camera-controls button:active {
+.adjuster-controls button:active {
     background: lightcyan;
 }
